requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
torch==2.1.0
torchvision==0.15.1
pillow==9.5.0
